requirements.txt
requests>=2.28.0
qrcode>=7.4.0
websocket-client>=1.6.0
